python-base/02-python-frame/fastapi/pyproject.toml
[project]
name = "fastapi-crud-demo"
version = "0.1.0"
description = "A simple CRUD demo with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
